# Optimization for input selection in case of huge number of UTXOs