--- a/tests/Bitcoin/InputSelectorTests.cpp
+++ b/tests/Bitcoin/InputSelectorTests.cpp
@@ -451,8 +451,8 @@ TEST(BitcoinInputSelector, ManyUtxos_900) {
     EXPECT_TRUE(verifySelectedUTXOs(selected, subset));
 }
 
-TEST(BitcoinInputSelector, ManyUtxos_4000_simple) {
-    const auto n = 4000;
+TEST(BitcoinInputSelector, ManyUtxos_5000_simple) {
+    const auto n = 5000;
     const auto byteFee = 10;
     std::vector<int64_t> values;
     uint64_t valueSum = 0;
@@ -461,22 +461,22 @@ TEST(BitcoinInputSelector, ManyUtxos_4000_simple) {
         values.push_back(val);
         valueSum += val;
     }
-    const uint64_t requestedAmount = valueSum / 8;
-    EXPECT_EQ(requestedAmount, 100'025'000);
+    const uint64_t requestedAmount = valueSum / 20;
+    EXPECT_EQ(requestedAmount, 62'512'500);
     auto utxos = buildTestUTXOs(values);
 
     auto selector = InputSelector<UTXO>(utxos);
     auto selected = selector.selectSimple(requestedAmount, byteFee);
 
-    // expected result: 1501 utxos, with the smaller amounts (except the very small dust ones)
+    // expected result: 1205 utxos, with the smaller amounts (except the very small dust ones)
     std::vector<int64_t> subset;
     uint64_t subsetSum = 0;
-    for (int i = 10; i < 1501+10; ++i) {
+    for (int i = 10; i < 1205+10; ++i) {
         const uint64_t val = (i + 1) * 100;
         subset.push_back(val);
         subsetSum += val;
     }
-    EXPECT_EQ(subset.size(), 1501);
-    EXPECT_EQ(subsetSum, 114'226'100);
+    EXPECT_EQ(subset.size(), 1205);
+    EXPECT_EQ(subsetSum, 73'866'500);
     EXPECT_TRUE(verifySelectedUTXOs(selected, subset));
 }