… tests/Bitcoin/TransactionPlanTests.cpp
… tests/Bitcoin/TxComparisonHelper.cpp