## Description

Optimizations for cases with large number of UTXOs:
- In cases with over 1000 UTXOs, use simplified selection strategy -- uisng a single pass over UTXOs (O(N) instead of O(N log N)).
- Always limit to max. 3000 the number of UTXOs used, potentially resulting in a lower transfer amount (even in maxAmount case).

Fixes #1639.

Some benchmarks:
- 2000 utxos, 601 selected:   master:  7.18s   PR: 4.63s  35% speedup
- 5000 utxos, maxamount:    master: 234.0s   PR: 42.9s  81% speedup

(wc part, on dev laptop, median of 5 measurements).

## Testing instructions

Relevant unit tests.

## Types of changes

<!--- What types of changes does your code introduce? Uncomment all the bullets that apply: -->

<!-- * Bug fix (non-breaking change which fixes an issue) -->

* Optimization (non-breaking change which does not affect functionality, but improves performance) -->

<!-- * New feature (non-breaking change which adds functionality) -->

<!-- * Breaking change (fix or feature that would cause existing functionality to change) -->

## Checklist

<!--- The following points should be used to indicate the progress of your PR.  Put an `x` in all the boxes that apply right now, and come back over time and check them off as you make progress.  If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [x] Prefix PR title with `[WIP]` if necessary.
- [x] Add tests to cover changes as needed.
- [x] Update documentation as needed.
- [x] I have read the [guidelines](https://developer.trustwallet.com/wallet-core/newblockchain#integration-criteria) for adding a new blockchain
- [x] If there is a related Issue, mention it in the description (e.g. Fixes #<issue_number> ).